[#18] Changes to genquery.py to accommodate msiGetMoreRows fix
Since genQueryOut structure's continueInx flag is (once again) an unambiguous
signal of when to terminate a loop paginating through the results of an iRODS
ICAT general query, the check of msiGetMoreRows's return code against
CAT_NO_ROWS_FOUND is now no longer necessary.

[irods/irods #4157] and this issue [#18] are both addressed by a modification to
msiGetMoreRows, as they apparently arose from the same cause: allowing a negative
value to persist in rei->status (on exit from msiGetMoreRows, when CAT_NO_ROWS_FOUND
was returned from rsGenQuery) actually prevented the update to the continueInx flag
from bubbling up to the microservice's caller, thus preventing its proper monitoring.

This commit contains requisite changes for the genquery.py module's main loop.
That loop, and the generator-style "row_iterator" function that contains it, are a
good example of how are microservices are/were traditionally used to iterate
through the query's results.

Also, the last call to msiGetMoreRows (with the aforementioned flag equal to zero) should
internally close the query and free resources associated with its use.  Although  not
normally apropos, it is possible that when both (1) general queries are in heavy or
frequent use within the lifetime of an iRODS server agent and (2) the writer of Python
rules is doing something to terminate pagination prematurely (say by 'break'-ing out of
the query-results iteration loop or otherwise forcing the garbage collection of the
iterator object prior to the its running the full course), something might have to be done
to mitigate memory and/or resource leaks.

The AUTO_CLOSE_QUERY attribute of the genquery module may optionally set to 'True' if desired.
This should have the effect of forcing the iteration to its end regardless of external
effects such as the iterator's being forced to exit prematurely; but please note the feature
is at present experimental.