[#49] PREP now supports the "return" keyword and REPF continuations.
Python rules can now return values back to the REPF via the "return"
keyword. This gives the PREP the ability to trigger continuations and
other actions (e.g. RULE_ENGINE_SKIP_OPERATION) in the REPF.

This change maintains the previous behavior. The PREP will not trigger
continuation by default. Rule implementers are required to return the
correct codes to achieve the behavior they want.

If the "return" keyword is not used, the PREP will not trigger
continuation and continue with the traditional flow.

- Added fmt library to build for completeness.
- Introduced new logger to the PREP.
- Added missing include to type_sequence.hpp.
- Resolved all warnings.