[#17] fix Query.copy: in python3, dict.items() no longer produces a list
As such, f( **(d1.items()+d2.items()) ) is no longer a valid construct in Python3,
and we must express it instead as f(**{**d1,**d2}). The meaning is of course the
same, and that is: in calling f, apply the keywords in d1 but overridden with the
values in d2 wherever there is overlap.