[#168] Add missing iRODS types and members
This change adds missing check_auth_credentials type.

Adds tcp_keepalive members to rodsEnv as well as connection pool timeout
member which was not added before.

Also added session_signature member of RcComm and session_props member of
RsComm.

-Wmissing-field-initializers is raised by a missing member of sockaddr_in
called sin_zero. This member has been removed in bwg2001-004. For more
information, see http://www.opengroup.org/platform/resolutions/bwg2001-many.html

Even though sin_zero must be included in the member list for sockaddr_in
for a complete member initialization list, we do not expose it as a
property for this type in the rule engine because sin_zero is not part
of the standard and may not be available in the future.

The missing member is included to silence the warning.