[#170] Make plugin interface functions static.
Before this change, it was possible for other rule engine plugins to
invoke the python rule engine plugin's interface functions. This is very
bad because it can lead to sibling plugins not operating correctly
and/or a broken server.

Marking the interface functions as static makes them private to the
implementation of the plugin. This means no other rule engine plugin
will be able to invoke the python rule engine plugin's interface
functions.